refactor: refactor package name from board to reference
